Added the vimrc file
This vimrc file is based off what I originally had. However, it has been
updated to use Vundle as the vim plugin manager. In addition, the list
of plugins can be found in this file.

Note that the status bar uses the newer, python version of Powerline.
This can be installed via http://lokaltog.github.com/powerline/

The font currently in use in both the terminal and GUI is PragmataPro.