Added select mode to powerline, solved eval problem with Ultisnips.
Previously, UltiSnips threw an error "Ultisnips: eval did not return a
valid python object". This was actually due to the gray powerline
configuration, which did not specify a color scheme for when in select
mode. Since UltiSnips moves into select mode when moving into fields, it
threw the error.

Now this is fixed, and a dark green "SELECT" mode indicator appears when
entering select mode.