Set up omni completion with SuperTab, use tab to expand.
This update allows for omni completion via tab in insert mode. In order
to make this compatible with UltiSnips, the key bindings for UltiSnips
were changed back the defaults. The exception for UltiSnips is the list
snippets key was changed to <c-k> in insert mode because <c-tab> (the
default) does not work by default in many terminals.