Go code files are automatically formatted on read, spaces, width 2.
Go code is automatically formatted with tabs instead of spaces, and a
width of 2. This is performed on the read of a file, so saving the file
to disk will not show the changes in vim. To do so, reload the file with
:e.

The reason for not performing the edit on write was because doing so
loses the place that the user was at, bumping them up to the top of the
file.